update db update script remove inlinecallbacks
