--- alembic/versions/b2626174a3af_add_dmhy_and_acg_rip_in_bangumi.py
@@ -1,30 +1,0 @@
-"""Add dmhy and acg_rip in Bangumi
-
-Revision ID: b2626174a3af
-Revises: 
-Create Date: 2016-08-20 20:47:22.371886
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = 'b2626174a3af'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('bangumi', sa.Column('acg_rip', sa.TEXT(), nullable=True))
-    op.add_column('bangumi', sa.Column('dmhy', sa.TEXT(), nullable=True))
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('bangumi', 'dmhy')
-    op.drop_column('bangumi', 'acg_rip')
-    ### end Alembic commands ###
