+++ alembic/versions/f798deb267f8_add_some_new_field_in_bangumi_and_new_.py
@@ -1,0 +1,38 @@
+"""add some new field in Bangumi and new talbe feed
+
+Revision ID: f798deb267f8
+Revises: 
+Create Date: 2016-09-11 11:32:48.365493
+
+"""
+
+# revision identifiers, used by Alembic.
+revision = 'f798deb267f8'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+def upgrade():
+    ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('feed',
+    sa.Column('episode_id', postgresql.UUID(as_uuid=True), nullable=False),
+    sa.Column('bangumi_id', postgresql.UUID(as_uuid=True), nullable=False),
+    sa.Column('download_url', sa.TEXT(), nullable=False),
+    sa.Column('torrent_file_id', postgresql.UUID(as_uuid=True), nullable=True),
+    sa.PrimaryKeyConstraint('episode_id')
+    )
+    op.add_column(u'bangumi', sa.Column('acg_rip', sa.TEXT(), nullable=True))
+    op.add_column(u'bangumi', sa.Column('dmhy', sa.TEXT(), nullable=True))
+    ### end Alembic commands ###
+
+
+def downgrade():
+    ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column(u'bangumi', 'dmhy')
+    op.drop_column(u'bangumi', 'acg_rip')
+    op.drop_table('feed')
+    ### end Alembic commands ###
